requirements.txt
discord
requests
Pillow
discord-py-slash-command
discord-py-interactions
beautifulsoup4
